+++ continuousintegration/docker/ctafrontend/cc7/opt/run/bin/taped.sh
@@ -51,6 +51,7 @@ echo ${DATABASEURL} > /etc/cta/cta-catalogue.conf
   echo "taped BufferCount 200" >> /etc/cta/cta-taped.conf
   echo "taped MountCriteria 2000000, 100" >> /etc/cta/cta-taped.conf
   echo "ObjectStore BackendPath $OBJECTSTOREURL" >> /etc/cta/cta-taped.conf
+  echo "taped UseEncryption no" >> /etc/cta/cta-taped.conf
   echo "${tpconfig}" > /etc/cta/TPCONFIG
 
 ####