+++ tapeserver/castor/tape/tapeserver/daemon/DataTransferConfig.cpp
@@ -33,6 +33,7 @@ castor::tape::tapeserver::daemon::DataTransferConfig::DataTransferConfig()
   nbDiskThreads(0),
   useLbp(false),
   useRAO(false),
+  useEncryption(true),
   externalEncryptionKeyScript(""),
   fetchEosFreeSpaceScript(""){}