+++ tapeserver/castor/tape/tapeserver/daemon/DataTransferSession.cpp
@@ -248,7 +248,7 @@ castor::tape::tapeserver::daemon::Session::EndOfSessionAction
 
     TapeReadSingleThread trst(*drive, m_mc, tsr, m_volInfo,
         m_castorConf.bulkRequestRecallMaxFiles,m_capUtils,rwd,lc,rrp,
-        m_castorConf.useLbp, m_castorConf.useRAO, m_castorConf.externalEncryptionKeyScript,*retrieveMount, m_castorConf.tapeLoadTimeout);
+        m_castorConf.useLbp, m_castorConf.useRAO, m_castorConf.useEncryption, m_castorConf.externalEncryptionKeyScript,*retrieveMount, m_castorConf.tapeLoadTimeout);
     DiskWriteThreadPool dwtp(m_castorConf.nbDiskThreads,
         rrp,
         rwd,
@@ -379,6 +379,7 @@ castor::tape::tapeserver::daemon::Session::EndOfSessionAction
         m_castorConf.maxFilesBeforeFlush,
         m_castorConf.maxBytesBeforeFlush,
         m_castorConf.useLbp,
+        m_castorConf.useEncryption,
         m_castorConf.externalEncryptionKeyScript,
         *archiveMount,
         m_castorConf.tapeLoadTimeout);